issue-992: extract github helper modules to crate
