+++ crates/tau-coding-agent/Cargo.toml
@@ -34,6 +34,7 @@ tau-diagnostics = { path = "../tau-diagnostics" }
 tau-release-channel = { path = "../tau-release-channel" }
 tau-ops = { path = "../tau-ops" }
 tau-memory = { path = "../tau-memory" }
+tau-github-issues = { path = "../tau-github-issues" }
 tau-agent-core = { path = "../tau-agent-core" }
 tau-ai = { path = "../tau-ai" }
 reqwest.workspace = true
